refactor(behaviors): remove skill related implementations and cleanup
