+++ src/plugins/behaviors/src/components.rs
@@ -1,30 +1,4 @@
 pub(crate) mod face_target;
-pub(crate) mod fix_points;
-pub(crate) mod ground_target;
+pub(crate) mod facing;
 pub(crate) mod movement;
 pub(crate) mod movement_definition;
-pub(crate) mod set_motion_forward;
-pub(crate) mod skill_behavior;
-pub(crate) mod when_traveled_insert;
-
-use bevy::prelude::*;
-use common::traits::handles_orientation::Face;
-use macros::SavableComponent;
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, PartialEq, Clone, Copy)]
-pub(crate) struct Always;
-
-#[derive(Debug, PartialEq, Clone, Copy)]
-pub(crate) struct Once;
-
-#[derive(Component, SavableComponent, Debug, PartialEq, Clone, Serialize, Deserialize)]
-#[require(CanFace)]
-pub struct SetFaceOverride(pub Face);
-
-#[derive(Component, SavableComponent, Debug, PartialEq, Clone, Serialize, Deserialize)]
-#[require(CanFace)]
-pub struct SetFace(pub Face);
-
-#[derive(Component, Debug, PartialEq, Default)]
-pub struct CanFace;
