--- src/plugins/behaviors/src/components/skill_behavior/skill_contact.rs
@@ -1,54 +1,0 @@
-pub(crate) mod dto;
-
-use super::SimplePrefab;
-use crate::components::skill_behavior::{FaultyColliderShape, skill_contact::dto::SkillContactDto};
-use bevy::prelude::*;
-use common::traits::{
-	handles_physics::HandlesPhysicalObjects,
-	handles_skill_behaviors::{Contact, ContactShape, Motion},
-	load_asset::LoadAsset,
-	prefab::{Prefab, PrefabEntityCommands},
-};
-use macros::SavableComponent;
-
-#[derive(Component, SavableComponent, Debug, Clone)]
-#[require(Visibility, Transform)]
-#[savable_component(dto = SkillContactDto)]
-pub struct SkillContact {
-	pub(crate) created_from: CreatedFrom,
-	pub(crate) shape: ContactShape,
-	pub(crate) motion: Motion,
-}
-
-#[derive(Debug, PartialEq, Clone, Copy)]
-pub(crate) enum CreatedFrom {
-	Contact,
-	Save,
-}
-
-impl From<Contact> for SkillContact {
-	fn from(Contact { shape, motion }: Contact) -> Self {
-		Self {
-			created_from: CreatedFrom::Contact,
-			shape,
-			motion,
-		}
-	}
-}
-
-impl<TPhysics> Prefab<TPhysics> for SkillContact
-where
-	TPhysics: HandlesPhysicalObjects,
-{
-	type TError = FaultyColliderShape;
-	fn insert_prefab_components(
-		&self,
-		entity: &mut impl PrefabEntityCommands,
-		_: &mut impl LoadAsset,
-	) -> Result<(), FaultyColliderShape> {
-		let created_from = self.created_from;
-
-		let Ok(()) = self.motion.prefab::<TPhysics>(entity, created_from);
-		self.shape.prefab::<TPhysics>(entity, Vec3::ZERO)
-	}
-}
