+++ src/plugins/behaviors/src/system_param.rs
@@ -1,3 +1,2 @@
 pub(crate) mod face_param;
 pub(crate) mod movement_param;
-pub(crate) mod skill_spawner;
