+++ src/plugins/behaviors/src/systems.rs
@@ -1,3 +1,2 @@
 pub(crate) mod face;
 pub(crate) mod movement_definition;
-pub(crate) mod update_skill_contact_range;
