--- src/plugins/behaviors/src/traits/has_filter.rs
@@ -1,5 +1,0 @@
-use bevy::ecs::query::QueryFilter;
-
-pub(crate) trait HasFilter {
-	type TFilter: QueryFilter;
-}
